requirements/base.txt
# General
# ------------------------------------------------------------------------------
pytz==2019.1  # https://github.com/stub42/pytz

# Django
# ------------------------------------------------------------------------------
django==2.2.10  # pyup: < 2.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-allauth==0.39.1  # https://github.com/pennersr/django-allauth
django-filter==2.1.0  # https://github.com/carltongibson/django-filter

# Django REST Framework
# ------------------------------------------------------------------------------
djangorestframework==3.9.2  # https://github.com/encode/django-rest-framework

# Swagger
# ------------------------------------------------------------------------------
drf-yasg==1.15.0
inflection==0.3.1
ruamel.yaml==0.15.97

# REST-JWT
# ------------------------------------------------------------------------------
PyJWT==1.7.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
